+++ plugin/controller.js
@@ -1,14 +1,15 @@
-var async = require('async');
+const async = require('async');
 
-var database = require('./database'),
-    logger   = require('./logger'),
-    rules    = require('./rules');
+const database = require('./database'),
+      logger   = require('./logger'),
+      rules    = require('./rules'),
+      Utils    = require('./utils');
 
 (function (Controller) {
 
     Controller.createRule = function (payload, done) {
         async.series([
-            async.apply(database.createRule, payloadToRule(payload)),
+            async.apply(database.createRule, Utils.payloadToRule(payload)),
             async.apply(rules.invalidate)
         ], done);
     };
@@ -31,17 +32,18 @@ var database = require('./database'),
     };
 
     Controller.installDefaultRules = function (done) {
-        var data = require('../data/default-rules');
-        var installed = [];
+        let data = require('../data/default-rules');
+        let installed = [];
 
         async.waterfall([
             async.apply(database.getRules),
             function (rules, callback) {
-                var toInstall = [], i = 0, len = data.rules.length, defaultRule;
+                let toInstall = [], i = 0, len = data.rules.length, defaultRule;
 
                 for (i; i < len; ++i) {
                     defaultRule = data.rules[i];
-                    if (isInList('name', defaultRule.name, rules)) {
+
+                    if (Utils.isInList('name', defaultRule.name, rules)) {
                         logger.log('verbose', 'Rule "%s" is skipped. Reason: already installed', defaultRule.displayName);
                     } else {
                         toInstall.push(defaultRule);
@@ -52,7 +54,7 @@ var database = require('./database'),
             },
             function (install, callback) {
                 async.eachSeries(install, function (defaultRule, next) {
-                    database.createRule(payloadToRule(defaultRule), function (error) {
+                    database.createRule(Utils.payloadToRule(defaultRule), function (error) {
                         if (error) {
                             return next(error);
                         }
@@ -109,43 +111,16 @@ var database = require('./database'),
 
     Controller.saveRule = function (rule, done) {
         async.series({
-            save: async.apply(database.updateRule, rule.rid, payloadToRule(rule)),
+            save: async.apply(database.updateRule, rule.rid, Utils.payloadToRule(rule)),
             rule: async.apply(database.getRule, rule.rid),
             sync: async.apply(rules.invalidate)
         }, function (error, results) {
             if (error) {
                 return done(error);
             }
+            
             done(null, results.rule)
         });
     };
-
-    function isInList(field, value, list) {
-        var i = 0, len = list.length, listItem;
-
-        for (i; i < len; ++i) {
-            listItem = list[i];
-
-            if (listItem[field] === value) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    function payloadToRule(payload) {
-        var rule = {};
-
-        // TODO Validation?
-
-        rule.name = payload.name;
-        rule.displayName = payload.displayName;
-        rule.regex = payload.regex;
-        rule.replacement = payload.replacement;
-        rule.icon = payload.icon || 'fa-cogs';
-
-        return rule;
-    }
 
 })(module.exports);
